views/flight: Move schemas into "schemas" package
