+++ skylines/frontend/views/flight.py
@@ -18,12 +18,12 @@ from skylines.lib.geo import METERS_PER_DEGREE
 from skylines.lib.geoid import egm96_height
 from skylines.lib.vary import vary
 from skylines.model import (
-    User, Flight, FlightPhase, Location, FlightComment,
+    User, Flight, Location, FlightComment,
     Notification, Event, FlightMeetings, AircraftModel,
 )
 from skylines.model.event import create_flight_comment_notifications
 from skylines.model.flight import get_elevations_for_flight
-from skylines.schemas import fields, FlightSchema, FlightCommentSchema, Schema, ValidationError
+from skylines.schemas import fields, FlightSchema, FlightCommentSchema, FlightPhaseSchema, ContestLegSchema, Schema, ValidationError
 from skylines.worker import tasks
 from redis.exceptions import ConnectionError
 
@@ -150,46 +150,6 @@ def _get_contest_traces(flight):
                                    times=xcsoar.encode(times, method="signed")))
 
     return contest_traces
-
-
-PHASETYPE_IDS = {
-    FlightPhase.PT_POWERED: u'powered',
-    FlightPhase.PT_CIRCLING: u'circling',
-    FlightPhase.PT_CRUISE: u'cruise',
-}
-
-
-CIRCDIR_IDS = {
-    FlightPhase.CD_LEFT: u'left',
-    FlightPhase.CD_MIXED: u'mixed',
-    FlightPhase.CD_RIGHT: u'right',
-    FlightPhase.CD_TOTAL: u'total',
-}
-
-
-class FlightPhaseSchema(Schema):
-    circlingDirection = fields.Function(lambda phase: CIRCDIR_IDS.get(phase.circling_direction))
-    type = fields.Function(lambda phase: PHASETYPE_IDS.get(phase.phase_type))
-    secondsOfDay = fields.Int(attribute='seconds_of_day')
-    startTime = fields.DateTime(attribute='start_time')
-    duration = fields.TimeDelta()
-    altDiff = fields.Float(attribute='alt_diff')
-    distance = fields.Float()
-    vario = fields.Float()
-    speed = fields.Float()
-    glideRate = fields.Float(attribute='glide_rate')
-    fraction = fields.Float()
-    count = fields.Int()
-
-
-class ContestLegSchema(Schema):
-    distance = fields.Float()
-    duration = fields.TimeDelta()
-    start = fields.Int(attribute='seconds_of_day')
-    climbDuration = fields.TimeDelta(attribute='climb_duration')
-    climbHeight = fields.Float(attribute='climb_height')
-    cruiseDistance = fields.Float(attribute='cruise_distance')
-    cruiseHeight = fields.Float(attribute='cruise_height')
 
 
 def mark_flight_notifications_read(flight):
